+++ src/aktualizr_secondary/CMakeLists.txt
@@ -60,7 +60,7 @@ include(AddAktualizrTest)
 list(INSERT TEST_LIBS 0 aktualizr_secondary_lib)
 
 add_aktualizr_test(NAME aktualizr_secondary_config
-    SOURCES aktualizr_secondary_config_test.cc PROJECT_WORKING_DIRECTORY LIBRARIES aktualizr_secondary_lib)
+                   SOURCES aktualizr_secondary_config_test.cc PROJECT_WORKING_DIRECTORY LIBRARIES aktualizr_secondary_lib)
 
 add_aktualizr_test(NAME aktualizr_secondary_update
                    SOURCES update_test.cc
@@ -81,8 +81,17 @@ if(BUILD_OSTREE)
                        SOURCES uptane_test.cc
                        LIBRARIES aktualizr_secondary_lib uptane_generator_lib virtual_secondary $<TARGET_OBJECTS:campaign> $<TARGET_OBJECTS:primary> $<TARGET_OBJECTS:http>
                        ARGS ${PROJECT_BINARY_DIR}/ostree_repo PROJECT_WORKING_DIRECTORY)
+
+    add_aktualizr_test(NAME aktualizr_secondary_uptane_verification_ostree
+                       SOURCES uptane_verification_ostree_test.cc
+                       ARGS ${PROJECT_BINARY_DIR}/ostree_repo
+                       LIBRARIES aktualizr_secondary_lib uptane_generator_lib $<TARGET_OBJECTS:campaign>
+                       PROJECT_WORKING_DIRECTORY)
+
+    set_target_properties(t_aktualizr_secondary_uptane_verification_ostree PROPERTIES LINK_FLAGS -Wl,--export-dynamic)
+    target_link_libraries(t_aktualizr_secondary_uptane_verification_ostree aktualizr_secondary_lib uptane_generator_lib)
 else(BUILD_OSTREE)
-    list(APPEND TEST_SOURCES uptane_verification_test.cc uptane_test.cc)
+    list(APPEND TEST_SOURCES uptane_verification_test.cc uptane_test.cc uptane_verification_ostree_test.cc)
 endif(BUILD_OSTREE)
 
 # test running the executable with command line option --help
