+++ src/aktualizr_secondary/aktualizr_secondary_config_test.cc
@@ -13,8 +13,11 @@ TEST(aktualizr_secondary_config, config_toml_parsing) {
   AktualizrSecondaryConfig conf("tests/config/aktualizr_secondary.toml");
 
   EXPECT_EQ(conf.network.port, 9031);
-
+#ifdef BUILD_OSTREE
   EXPECT_EQ(conf.pacman.type, PackageManager::kOstree);
+#else
+  EXPECT_EQ(conf.pacman.type, PackageManager::kNone);
+#endif
   EXPECT_EQ(conf.pacman.os, std::string("testos"));
   EXPECT_EQ(conf.pacman.sysroot, boost::filesystem::path("testsysroot"));
   EXPECT_EQ(conf.pacman.ostree_server, std::string("test_server"));
