+++ src/libaktualizr/uptane/directorrepository.h
@@ -21,16 +21,18 @@ class DirectorRepository : public RepositoryCommon {
     return targets.getTargets(ecu_id, hw_id);
   }
   const std::string& getCorrelationId() const { return targets.correlation_id(); }
-  bool targetsExpired() const;
+
   bool checkMetaOffline(INvStorage& storage);
   void dropTargets(INvStorage& storage);
 
   Exception getLastException() const { return last_exception; }
   bool updateMeta(INvStorage& storage, const IMetadataFetcher& fetcher) override;
+  bool matchTargetsWithImageTargets(const Uptane::Targets& image_targets) const;
 
  private:
   void resetMeta();
   bool usePreviousTargets() const;
+  bool targetsExpired() const;
 
  private:
   FRIEND_TEST(Director, EmptyTargets);
